+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * CHECK: check websites for favicons, if favicon is currently empty
 * CLEANUP: recheck favicon every day if still existing and delete from station if not
+* CONFIG: regular automatic load/download of config
 
 ### Changed
 * CLEANUP: add station change entries into the database for language and favicon updates